TSQL: Create coverage tests for TSQL -> Databricks functions  (#415)
As we now have translation capability for them, we start out by
examining the TSQL functions and providing the equivalent Databricks SQL
for them.

- Some functions, mainly the common ANSI defined SQL functions are
directly equivalent.
  - Some have the same name but mean different things.
  - Some functions have equivalent ways to achieve the same thing.
- Some functions will likely be unsupportable without changes to
Databricks SQL.
- Some functions may be unsupportable due to fundamental differences
between TSQL and Databricks SQL engines/models.
  
Files are supplied under the tests/resources/functional/tsql directory.
The TSQL and Databricks SQL equivlent are provided and the conversion is
documented before the TSQL is introduced as we may be able to automate
documentation to a certain degree using this commentary.

---------

Co-authored-by: Valentin Kasas <valentin.kasas@xebia.com>