Add Oracle ojdbc8 dependent library during reconcile Installation (#474)
As part of the `reconcile` process, we may need to pull the data from
the `Oracle` Data source for `reconciliation` between `Oracle` and
`Databricks`. It requires `ojdbc8.jar` on the CLASSPATH.

This PR handles:
- the installation of `com.oracle.database.jdbc:ojdbc8:23.4.0.24.05` jar
as a dependent library in the `Remorph_Reconciliation_Job` Workflow.
- Updated the driver class to `oracle.jdbc.driver.OracleDriver` instead
of `oracle` in `JDBCReaderMixin`

<img width="770" alt="image"
src="https://github.com/databrickslabs/remorph/assets/84904851/ad04231c-d627-4b07-b77b-ef01fd99a06c">