Generic stored procedure parsing (#1047)
Here we create an ANTLR grammar library mechanism for including
grammars, homogonize certain high level rule names, and implement a
shared grammar for stored procedures.

The shared grammar caters for TSQL and Snowflake stored procedures, but
is intended to be expanded for all future SQL dialects as the standard
SQL stored procedures are all very similar and dialect specific
extensions such as LANGUAGE xyz in Snowflake do not create unresolvable
ambiguities.

The lexers for Snowflake and TSQL have been redesigned to share many
common tokens, leaving only a few for the specific lexers, where they
would take too much time to merge at the moment. But the intention is to
eventually have a common lexer for all dialects, once the over-specified
tokens (such as option names) are replaced with `id` or `genericOption`.

This PR is essentially partially finished as the complete stored
procedure grammar is not yet defined and the rules for TSQL and
Snowflake are yet to be combined. This is left for a future PR.

NB: The ANTLR grammar linter will have to be upgraded to cater for
library grammars and to process includes as it is currently only able to
process a single grammar. For now, the linting build step is changed to
continue on error.