Fix Snowflake precedence of `INTERSECT` with respect to other set operations (#1288)
This PR updates the Snowflake processing so that the precedence of
`INTERSECT` is handled correctly. At present the set operations
(`UNION`, `EXCEPT` and `INTERSECT`) are all handled left-to-right with
equal precedence. However this is incorrect, Snowflake semantics
are[^1], from highest to lowest:
1. Parenthesis
2. `INTERSECT`
3. `UNION [ALL]` and `MINUS`/`EXCEPT`

At the same level operators are processed from left to right.

Resolves #1255.

Tests coverage:
 - Unit tests.
 - Transpiler tests.
 - Functional tests.

Incidental changes include factoring out some of the tests that we run
against TSQL so that the same tests can run against Snowflake.

[^1]: Reference:
https://docs.snowflake.com/en/sql-reference/operators-query